Aruudy/requirements.txt
python-levenshtein
lang-trans
